Highlight jumps in dex code
Quick and dirty text based implementation. I need to rework
the parser to create a proper structure for the parsed code
so the jumps can be highlighted from the source *and* from
the destination.

Properly structured code results would also allows to
highlight the disassembly nicely, and to offer a UI to
navigate through classes and methods.