Launch Some Task in Parallel to Speed Things Up
`javap` and `dexdump` can be launched in the background while we continue processing.

I changed the semantics of Status to show what has already been done, rather that's being done now. This is because now, things can happen at the same time. For example:

`Creating Java ByteCode` & `Creating Dex Dump` would be set at almost the same time and we wouldn't see one of them.

Now that we're running some updates in the background, we could have race conditions will leave the status bar with a wrong message, for example, if the DEXDUMP process ends after the OATDUMP, we end up with a wrong status.

I added a ProgressUpdater class that helps manage this and reports inconsistencies to the console if `KOTLIN_EXPLORER_LOG` is set.

Note that I'm tracking all spawned jobs, so I can join them at the end before sending the `Ready` status.